requirements.txt
Django==4.2.11
mysqlclient==2.1.0
djangorestframework==3.13.1
django-cors-headers==4.0.0
django-allauth==0.48.0
django-rest-auth==0.9.5
django-allauth-socialaccount==0.48.0
django-countries==7.3.2
PyJWT==2.3.0